+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Added
 - Setting for font family to use when Use mono font is active.
 - Setting to set the opacity levels for when a note is transparent or opaque.
+- Settings for default note size.
 - 
 
 ### Changed
@@ -23,6 +24,7 @@
 ### Fixed
 - Crash in cases when clipboard reports having text but has issue retrieving it.
 - 
+
 
 ## v1.12.1 - 21/11/2024